Updated UVI Calls with new Network Code
Updated a large number of UVI calls for Tree traversal to use different
networking logic.  The new logic is easier to read to allow other
developers a better understanding at first glance of what is going on.
Additionally the new code should help alleviate some of the issues that
the old networking code wasn't able too.