Modified the aggregation query textbox properties
Modified the aggregations query textbox to be able to handle multiline and
accept the Enter key to create a new line.  With all these attributes set
the textbox still isn't creating a new line correctly.  Still a work in
progress but these properties settings will be required for the future.