Addressed Issue #69
The issue is an out of memory issue when attempting to build the aoi
string for the web request.  A try/catch block has been added that will
catch the out of memory error and inform the user to select fewer
features.  The go button is also re-enabled in the catch block.